Fix github action ci

+++ internal/service/subscribe_test.go
@@ -4,6 +4,7 @@ import (
 	"github.com/gorilla/websocket"
 	"net/http"
 	"os"
+	"strings"
 	"syscall"
 	"testing"
 )
@@ -34,11 +35,13 @@ func (t *TestConn) Close() {
 }
 
 func (t *TestConn) ReadMessage() (messageType int, message []byte, err error) {
-	_ = syscall.Kill(syscall.Getpid(), syscall.SIGINT)
 	return
 }
 
 func (t *TestConn) WriteMessage(messageType int, data []byte) error {
+	if strings.EqualFold(string(data), `{"id":4,"method":"state_subscribeStorage","params":[["0x481e203dcea218263e3a96ca9e4b193857c875e4cff74148e4628f264b974c80"]],"jsonrpc":"2.0"}`) {
+		_ = syscall.Kill(syscall.Getpid(), syscall.SIGINT)
+	}
 	return nil
 }
